refactor: useStage.ts
