+++ src/lib/constants.ts
@@ -1,0 +1,2 @@
+export const CANVAS_WIDTH = 1920;
+export const CANVAS_HEIGHT = 1080;
